Clean up Uebung 9 StudentTests

+++ Uebungsblatt_09/H09_Stollenwerk_Vincent/src/H1/StudentTests.java
@@ -14,7 +14,7 @@ class StudentTests {
 
         intListInitialization(linkedList);
 
-        assertEquals(7, linkedList.search(integer -> integer == 7));
+        assertEquals(7, (int) linkedList.search(integer -> integer == 7));
 
         assertEquals("7 0 1 2 3 4 5 6 8 9 ", linkedList.toString());
     }
@@ -25,11 +25,11 @@ class StudentTests {
 
         intListInitialization(linkedList);
 
-        assertEquals(1, linkedList.search(integer -> integer == 1));
+        assertEquals(1, (int) linkedList.search(integer -> integer == 1));
 
         assertEquals("1 0 2 3 4 5 6 7 8 9 ", linkedList.toString());
 
-        assertEquals(7, linkedList.search(integer -> integer == 7));
+        assertEquals(7, (int) linkedList.search(integer -> integer == 7));
 
         assertEquals("1 0 2 3 4 5 7 6 8 9 ", linkedList.toString());
     }
@@ -40,14 +40,14 @@ class StudentTests {
 
         intListInitialization(linkedList);
 
-        assertEquals(1, linkedList.search(integer -> integer == 1));
+        assertEquals(1, (int) linkedList.search(integer -> integer == 1));
         assertEquals(1, linkedList.head.counter);
         assertEquals("1 0 2 3 4 5 6 7 8 9 ", linkedList.toString());
 
-        assertEquals(7, linkedList.search(integer -> integer == 7));
+        assertEquals(7, (int) linkedList.search(integer -> integer == 7));
         assertEquals("1 7 0 2 3 4 5 6 8 9 ", linkedList.toString());
 
-        assertEquals(7, linkedList.search(integer -> integer == 7));
+        assertEquals(7, (int) linkedList.search(integer -> integer == 7));
         assertEquals("7 1 0 2 3 4 5 6 8 9 ", linkedList.toString());
     }
